# Optimize TagHelper serialization & deserialization