--- a/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/TagHelperSerializationBenchmark.cs
+++ b/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/TagHelperSerializationBenchmark.cs
@@ -3,13 +3,17 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
+using System.Linq;
 using System.Text;
 using BenchmarkDotNet.Attributes;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.CodeAnalysis.Razor.Serialization;
 using Microsoft.VisualStudio.LanguageServices.Razor.Serialization;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Serialization;
 
 namespace Microsoft.AspNetCore.Razor.Performance
 {
@@ -27,28 +31,60 @@ namespace Microsoft.AspNetCore.Razor.Performance
 
             var tagHelperFilePath = Path.Combine(current.FullName, "taghelpers.json");
             _tagHelperBuffer = File.ReadAllBytes(tagHelperFilePath);
+
+            // Deserialize from json file.
+            DefaultSerializer = new JsonSerializer();
+            DefaultSerializer.Converters.Add(new TagHelperDescriptorJsonConverter());
+            using (var stream = new MemoryStream(_tagHelperBuffer))
+            using (var reader = new JsonTextReader(new StreamReader(stream)))
+            {
+                DefaultTagHelpers = DefaultSerializer.Deserialize<IReadOnlyList<TagHelperDescriptor>>(reader);
+            }
+        }
+
+        public IReadOnlyList<TagHelperDescriptor> DefaultTagHelpers { get; set; }
+
+        public JsonSerializer DefaultSerializer { get; set; }
+
+        [Benchmark(Description = "Razor TagHelper Roundtrip Serialization")]
+        public void TagHelper_Serialization_RoundTrip()
+        {
+            // Serialize back to json.
+            MemoryStream originalStream;
+            using (originalStream = new MemoryStream())
+            using (var writer = new StreamWriter(originalStream, Encoding.UTF8, bufferSize: 4096))
+            {
+                DefaultSerializer.Serialize(writer, DefaultTagHelpers);
+            }
+
+            IReadOnlyList<TagHelperDescriptor> reDeserializedTagHelpers;
+            var stream = new MemoryStream(originalStream.GetBuffer());
+            using (stream)
+            using (var reader = new JsonTextReader(new StreamReader(stream)))
+            {
+                reDeserializedTagHelpers = DefaultSerializer.Deserialize<IReadOnlyList<TagHelperDescriptor>>(reader);
+            }
         }
 
         [Benchmark(Description = "Razor TagHelper Serialization")]
-        public void TagHelper_Serialization_RoundTrip()
+        public void TagHelper_Serialization()
         {
-            var serializer = new JsonSerializer();
-            serializer.Converters.Add(new RazorDiagnosticJsonConverter());
-            serializer.Converters.Add(new TagHelperDescriptorJsonConverter());
+            using (var stream = new MemoryStream())
+            using (var writer = new StreamWriter(stream, Encoding.UTF8, bufferSize: 4096))
+            {
+                DefaultSerializer.Serialize(writer, DefaultTagHelpers);
+            }
+        }
 
+        [Benchmark(Description = "Razor TagHelper Deserialization")]
+        public void TagHelper_Deserialization()
+        {
             // Deserialize from json file.
             IReadOnlyList<TagHelperDescriptor> tagHelpers;
             using (var stream = new MemoryStream(_tagHelperBuffer))
             using (var reader = new JsonTextReader(new StreamReader(stream)))
             {
-                tagHelpers = serializer.Deserialize<IReadOnlyList<TagHelperDescriptor>>(reader);
-            }
-
-            // Serialize back to json.
-            using (var stream = new MemoryStream())
-            using (var writer = new StreamWriter(stream, Encoding.UTF8, bufferSize: 4096))
-            {
-                serializer.Serialize(writer, tagHelpers);
+                tagHelpers = DefaultSerializer.Deserialize<IReadOnlyList<TagHelperDescriptor>>(reader);
             }
         }
     }