--- a/src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DefaultRazorProjectChangePublisher.cs
+++ b/src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/DefaultRazorProjectChangePublisher.cs
@@ -32,10 +32,7 @@ namespace Microsoft.VisualStudio.LanguageServerClient.Razor
         private readonly Dictionary<string, ProjectSnapshot> _pendingProjectPublishes;
         private readonly object _publishLock;
 
-        private readonly JsonSerializer _serializer = new JsonSerializer()
-        {
-            Formatting = Formatting.Indented
-        };
+        private readonly JsonSerializer _serializer = new JsonSerializer();
 
         private ProjectSnapshotManagerBase _projectSnapshotManager;