… src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Performance/taghelpers.json
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer.Common/Serialization/TagHelperDescriptorJsonConverter.cs
… src/Razor/test/Microsoft.VisualStudio.LanguageServices.Razor.Test/test.taghelpers.json