- Removed a duplicate `TagHelperDescriptorJsonConverter` from language server binary in favor of platform agnostic variant.
- Completely re-worked TagHelper deserialization by not allocating `JObject`'s and instead reading directly from the `JsonReader`.
- Optimized TagHelper serialization by doing pre-checks for properties that are commonly provided and not serializing them if they'll default to the correct  value (i.e. not serializing a boolean `false` property).
- Changed our project.razor.json serialization behavior to serialize without spaces (unformatted).
- Updated the Benchmarks repo `README.md` and project file because it's what I had to do in order to run those tests.
- Updated Newtonsoft.Json to reflect what's delivered in Visual Studio
- Added a test that uses a test TagHelper input from a default Blazor server and ensures that the data can be serialized and deserialized into objects that are equivalent (round trip). Most other existing tests cover these changes but I wanted to take extra care that our serialization/deserialization worked as expected. Also offline I went ahead and ran the two TH json converters against each other and ensured that their outputs were identical.

## Benchmarks

#### Before

```
                                    Method |      Mean |     Error |    StdDev |   Op/s |    Gen 0 |   Gen 1 |   Gen 2 | Allocated |
------------------------------------------ |----------:|----------:|----------:|-------:|---------:|--------:|--------:|----------:|
 'Razor TagHelper Roundtrip Serialization' | 13.885 ms | 0.1916 ms | 0.1792 ms |  72.02 | 156.2500 | 93.7500 | 31.2500 |  10.66 MB |
           'Razor TagHelper Serialization' |  2.863 ms | 0.0567 ms | 0.1079 ms | 349.32 |  35.1563 | 35.1563 | 35.1563 |    1.2 MB |
         'Razor TagHelper Deserialization' | 10.989 ms | 0.0455 ms | 0.0404 ms |  91.00 | 125.0000 | 31.2500 |       - |   9.45 MB |
```

 #### After

```
                                    Method |     Mean |     Error |    StdDev |  Op/s |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
------------------------------------------ |---------:|----------:|----------:|------:|--------:|--------:|--------:|----------:|
 'Razor TagHelper Roundtrip Serialization' | 5.069 ms | 0.0930 ms | 0.0870 ms | 197.3 | 62.5000 | 54.6875 | 31.2500 |   3.87 MB |
           'Razor TagHelper Serialization' | 1.950 ms | 0.0389 ms | 0.0701 ms | 512.9 | 35.1563 | 33.2031 | 33.2031 |   1.11 MB |
         'Razor TagHelper Deserialization' | 4.355 ms | 0.0173 ms | 0.0161 ms | 229.6 | 39.0625 | 15.6250 |       - |   3.28 MB |
```

#### Result

|                                           | Op/s   | Allocated |
|-------------------------------------------|--------|-----------|
| 'Razor TagHelper Roundtrip Serialization' | +63%   | -64%      |
| 'Razor TagHelper Serialization'           | +31.9% | -7.5%     |
| 'Razor TagHelper Deserialization'         | +60%   | -65%      |

## project.razor.json size for a mostly default Blazor server

Before size: 817 KB
After size: 391 KB
Result: 53% reduced

Fixes dotnet/aspnetcore#23174